[storage] refactor: replace TransactionInfoIter with ContinousVersionIter
