+++ storage/aptosdb/src/event_store/mod.rs
@@ -6,6 +6,7 @@
 #![allow(unused)]
 
 use super::AptosDB;
+use crate::utils::iterators::EventsByVersionIter;
 use crate::{
     errors::AptosDbError,
     schema::{
@@ -73,13 +74,13 @@ impl EventStore {
         let mut iter = self.db.iter::<EventSchema>(Default::default())?;
         iter.seek(&start_version)?;
 
-        Ok(EventsByVersionIter {
-            inner: iter.peekable(),
-            expected_next_version: start_version,
-            end_version: start_version
+        Ok(EventsByVersionIter::new(
+            iter,
+            start_version,
+            start_version
                 .checked_add(num_versions as u64)
                 .ok_or_else(|| format_err!("Too many versions requested."))?,
-        })
+        ))
     }
 
     pub fn get_event_by_version_and_index(
@@ -557,46 +558,6 @@ struct EmptyReader;
 impl HashReader for EmptyReader {
     fn get(&self, _position: Position) -> Result<HashValue> {
         unreachable!()
-    }
-}
-
-pub struct EventsByVersionIter<'a> {
-    inner: Peekable<SchemaIterator<'a, EventSchema>>,
-    expected_next_version: Version,
-    end_version: Version,
-}
-
-impl<'a> EventsByVersionIter<'a> {
-    fn next_impl(&mut self) -> Result<Option<Vec<ContractEvent>>> {
-        if self.expected_next_version >= self.end_version {
-            return Ok(None);
-        }
-
-        let mut ret = Vec::new();
-        while let Some(res) = self.inner.peek() {
-            let ((version, _index), _event) = res
-                .as_ref()
-                .map_err(|e| format_err!("Hit error iterating events: {}", e))?;
-            if *version != self.expected_next_version {
-                break;
-            }
-            let ((_version, _index), event) =
-                self.inner.next().transpose()?.expect("Known to exist.");
-            ret.push(event);
-        }
-        self.expected_next_version = self
-            .expected_next_version
-            .checked_add(1)
-            .ok_or_else(|| format_err!("expected version overflowed."))?;
-        Ok(Some(ret))
-    }
-}
-
-impl<'a> Iterator for EventsByVersionIter<'a> {
-    type Item = Result<Vec<ContractEvent>>;
-
-    fn next(&mut self) -> Option<Self::Item> {
-        self.next_impl().transpose()
     }
 }
