+++ storage/aptosdb/src/pruner/ledger_store/ledger_store_pruner.rs
@@ -13,11 +13,12 @@ use crate::{
             transaction_store_pruner::TransactionStorePruner, write_set_pruner::WriteSetPruner,
         },
     },
+    pruner_utils,
     schema::{
         db_metadata::{DbMetadataKey, DbMetadataValue},
         transaction::TransactionSchema,
     },
-    utils, EventStore, StateStore, TransactionStore,
+    EventStore, StateStore, TransactionStore,
 };
 
 use aptos_logger::warn;
@@ -156,7 +157,7 @@ impl LedgerPruner {
         let target_version = 1; // The genesis version is 0. Delete [0,1) (exclusive)
         let max_version = 1; // We should only be pruning a single version
 
-        let ledger_pruner = utils::create_ledger_pruner(ledger_db, state_store);
+        let ledger_pruner = pruner_utils::create_ledger_pruner(ledger_db, state_store);
         ledger_pruner.set_target_version(target_version);
         ledger_pruner.prune_inner(max_version, db_batch)?;
