+++ storage/aptosdb/src/pruner/mod.rs
@@ -6,10 +6,10 @@ pub(crate) mod event_store;
 pub(crate) mod ledger_pruner_worker;
 pub(crate) mod ledger_store;
 pub(crate) mod pruner_manager;
+pub mod pruner_utils;
 pub(crate) mod state_pruner_worker;
 pub(crate) mod state_store;
 pub(crate) mod transaction_store;
-pub mod utils;
 
 // This module provides `Pruner` which manages a thread pruning old data in the background and is
 // meant to be triggered by other threads as they commit new data to the DB.
