+++ storage/aptosdb/src/pruner/pruner_utils.rs
